requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
Pillow==10.0.0
psycopg==3.1.12
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
gunicorn==21.2.0
requests==2.31.0
daphne==4.0.0


